Yom LaYabasha in a Yerushalmi melody
piyut.org.il says that this is a melody of Yerushalmi Sephardim. I'm not
sure whether to believe them, since this is sung by a couple of female
singers, and I can't find any other recordings by known sephardi
chazzanim. Anyhow, I've been listening to it a lot, and I like it, so
I'm documenting it.