The melody that I had for Al Bayit Zeh is actually Li Yah Li Yah.
The melody that I had for Al Bayit Zeh is actually Li Yah Li Yah, so I
changed the lyrics and the name. There seem to be 3 melodies floating
around for Al Bayit Zeh, and they are:

 * Li Yah Li Yah -- according to sephardichazzanut.com
 * Al Tirah Avdi Yaakov - according to pizmonim.org
 * Im Chacham Libecha Beni - according to Shir Ushbacha from Machon HaKtav.

Note that the lyrics for this song in Shir Ushbacha Hallel V'Zimrah (the
red book) are missing the first line. The recordings on pizmonim.org
include the first line. Shir Ushbacha from Machon HaKtav has it correct.