Incorporate the alternate verse meldoy into the main Daat Umzimah score.
In the main verse melody, change eighth-eighth to
dotted_eighth-sixteenth in one place. This doesn't help the fact that
"chochmah" seems to have 2 quarter notes on the "choch" part in most
recordings, but that makes the number of beats in the main verse melody
prime. I'll think about that problem separately.